Added Abstract network entity
